feat: Add template CMakeLists.txt and start HAL work
Here we go! This creates a library file that can be used by the Kernel
to do Hardware related things in an abstract way

Yes, this is basically identical to what WinNT does. I want to give this
a crack instead of having it all built into the Kernel Image in the same
subdir. This may end up being a huge mistake, but we'll see